+++ lua/custom/plugins/filetree.lua
@@ -1,34 +1,35 @@
 return {
-  {
-    'nvim-neo-tree/neo-tree.nvim',
-    version = '*',
-    dependencies = {
-      'nvim-lua/plenary.nvim',
-      'nvim-tree/nvim-web-devicons', -- not strictly required, but recommended
-      'MunifTanjim/nui.nvim',
-    },
-    lazy = false,
-    keys = {
-      { '\\', ':Neotree reveal<CR>', desc = 'NeoTree reveal', silent = true },
-    },
-    opts = {
-      filesystem = {
-        window = {
-          mappings = {
-            ['\\'] = 'close_window',
-          },
-        },
-      },
-    },
-    init = function()
-      -- 在离开 Tab 时自动关闭文件树边栏
-      vim.api.nvim_create_autocmd('TabLeave', {
-        callback = function()
-          vim.cmd 'Neotree close'
-        end,
-      })
-    end,
-  },
+  -- [[deprecated]]
+  -- {
+  --   'nvim-neo-tree/neo-tree.nvim',
+  --   version = '*',
+  --   dependencies = {
+  --     'nvim-lua/plenary.nvim',
+  --     'nvim-tree/nvim-web-devicons', -- not strictly required, but recommended
+  --     'MunifTanjim/nui.nvim',
+  --   },
+  --   lazy = false,
+  --   keys = {
+  --     { '\\', ':Neotree reveal<CR>', desc = 'NeoTree reveal', silent = true },
+  --   },
+  --   opts = {
+  --     filesystem = {
+  --       window = {
+  --         mappings = {
+  --           ['\\'] = 'close_window',
+  --         },
+  --       },
+  --     },
+  --   },
+  --   init = function()
+  --     -- 在离开 Tab 时自动关闭文件树边栏
+  --     vim.api.nvim_create_autocmd('TabLeave', {
+  --       callback = function()
+  --         vim.cmd 'Neotree close'
+  --       end,
+  --     })
+  --   end,
+  -- },
 
   {
     'stevearc/oil.nvim',