+++ lua/custom/plugins/filetree.lua
@@ -33,19 +33,17 @@ return {
 
   {
     'stevearc/oil.nvim',
+    ---@module 'oil'
+    ---@type oil.SetupOpts
+    opts = {
+      keymaps = {
+        ['<C-p>'] = { 'actions.preview', opts = { vertical = true, split = 'botright' } },
+      },
+    },
     -- Optional dependencies
-    -- dependencies = { 'echasnovski/mini.icons' },
-    dependencies = { 'nvim-tree/nvim-web-devicons' }, -- use if prefer nvim-web-devicons
-
-    config = function()
-      require('oil').setup {
-        view_options = {
-          show_hideen = true,
-        },
-      }
-
-      -- Open parent directory in current window.
-      vim.keymap.set('n', '-', '<CMD>Oil<CR>', { desc = 'Open parent directory' })
-    end,
+    dependencies = { { 'echasnovski/mini.icons', opts = {} } },
+    -- dependencies = { "nvim-tree/nvim-web-devicons" }, -- use if you prefer nvim-web-devicons
+    -- Lazy loading is not recommended because it is very tricky to make it work correctly in all situations.
+    lazy = false,
   },
 }