+++ CHANGELOG.md
@@ -2,18 +2,20 @@
 
 ## [Unreleased]
 
-### Added
+## [0.5.0] (personal) - 2025-05-19
+
+### Added (0.5.0)
 
 - mini.ai.
 - mini.surround.
 
-### Changed
+### Changed (0.5.0)
 
 - Replace onedark.nvim with tokyonight.nvim.
 - Replace lualine and bufferline with mini.nvim.
 - Make markdown conceallevel 2 active.
 
-### Removed
+### Removed (0.5.0)
 
 - Delete `<leader>e` keymap, `CTRL-W_d` is built-in.
 
@@ -70,6 +72,7 @@
 ## [0.1.0]
 
 [Unreleased]: https://github.com/nvim-lua/kickstart.nvim/compare/6ba2408cdf5eb7a0e4b62c7d6fab63b64dd720f6...master
+[0.5.0]: https://github.com/nvim-lua/kickstart.nvim/compare/6ba2408cdf5eb7a0e4b62c7d6fab63b64dd720f6...master
 [0.4.0]: https://github.com/nvim-lua/kickstart.nvim/compare/6ba2408cdf5eb7a0e4b62c7d6fab63b64dd720f6...master
 [0.3.0]: https://github.com/nvim-lua/kickstart.nvim/compare/6ba2408cdf5eb7a0e4b62c7d6fab63b64dd720f6...master
 [0.2.2]: https://github.com/nvim-lua/kickstart.nvim/compare/6ba2408cdf5eb7a0e4b62c7d6fab63b64dd720f6...master